requirements.txt
attrs==24.2.0
autopep8==2.3.1
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
colorama==0.4.6
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.20.0
google-auth==2.35.0
google-cloud-aiplatform==1.69.0
google-cloud-bigquery==3.26.0
google-cloud-bigquery-storage==2.26.0
google-cloud-core==2.4.1
google-cloud-functions==1.17.0
google-cloud-language==2.14.0
google-cloud-storage==2.18.2
google-cloud-translate==3.16.0
google-cloud-vision==3.7.4
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpc-google-iam-v1==0.13.1
grpcio==1.66.2
grpcio-status==1.66.2
idna==3.10
iniconfig==2.0.0
numpy==2.1.1
packaging==24.1
parameterized==0.9.0
pluggy==1.5.0
proto-plus==1.24.0
protobuf==5.28.2
py==1.11.0
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.1
pycodestyle==2.12.1
pyparsing==3.1.4
pytest==8.3.3
pytest-json-report==1.5.0
pytest-metadata==3.1.1
python-dateutil==2.9.0.post0
requests==2.32.3
rsa==4.9
six==1.16.0
smmap
toml==0.10.2
tomli==2.0.2
urllib3==2.2.3
